test: split browser runner.test.ts into per-scenario specs
runner.test.ts spawned ~11 independent browser runs in a single spec
file. Since --shard splits at spec-file granularity, this file couldn't
be sharded and had to be special-cased in CI. Split it into thematic
spec files so the sequencer can distribute the heavy runs, and drop the
viewport test that duplicated specs/viewport.test.ts.
